Move traits implementation of str to new mod
Also move FromStr trait
